Avoiding cyclic dependencies. But seems it wasnt needed.
